+++ database/repository/dex/dex.go
@@ -8,10 +8,10 @@ import (
 
 // Dex 레포지토리 인터페이스
 type DexRepository interface {
-	FindDexEventById(id int64) (res *entity.Event, err error)
-	FindDexDetailById(id int64) (res *entity.Detail, err error)
-	FindUserDexById(dexId int64, userId int64) (res int64, err error)
-	CreateUserDexById(dexId int64, userId int64) (err error)
+	FindDexEventById(id int) (res *entity.Event, err error)
+	FindDexDetailById(id int) (res *entity.Detail, err error)
+	FindUserDexById(dexId int, userId int) (res int, err error)
+	CreateUserDexById(dexId int, userId int) (err error)
 }
 
 type gormDexRepository struct {
@@ -22,7 +22,8 @@ func NewDexRepository(db *gorm.DB) DexRepository {
 	return &gormDexRepository{db}
 }
 
-func (g *gormDexRepository) FindDexEventById(id int64) (res *entity.Event, err error) {
+// [사건 내용 조회] 사건 id로 조회 select문
+func (g *gormDexRepository) FindDexEventById(id int) (res *entity.Event, err error) {
 	// 1. 쿼리작성
 	// select  e.name, e.level, d.define, d.outline, d.place, d.background, d.image_url
 	//   from "event" e, "detail" d
@@ -38,7 +39,8 @@ func (g *gormDexRepository) FindDexEventById(id int64) (res *entity.Event, err e
 	return res, nil
 }
 
-func (g *gormDexRepository) FindDexDetailById(id int64) (res *entity.Detail, err error) {
+// [사건 내용 조회] 사건 detail id로 조회 select문
+func (g *gormDexRepository) FindDexDetailById(id int) (res *entity.Detail, err error) {
 	// 1. 쿼리작성
 	// select  e.name, e.level, d.define, d.outline, d.place, d.background, d.image_url
 	//   from "event" e, "detail" d
@@ -54,7 +56,8 @@ func (g *gormDexRepository) FindDexDetailById(id int64) (res *entity.Detail, err
 	return res, nil
 }
 
-func (g *gormDexRepository) FindUserDexById(dexId int64, userId int64) (res int64, err error) {
+// [사용자 사건 수집 등록] 사건 id로 조회 select문 : 사건보유 여부 위함
+func (g *gormDexRepository) FindUserDexById(dexId int, userId int) (res int, err error) {
 	// 1. 쿼리작성
 	// select * from userdex where dex_id = 1 and user_id = 1
 
@@ -65,10 +68,11 @@ func (g *gormDexRepository) FindUserDexById(dexId int64, userId int64) (res int6
 	if tx.Error != nil {
 		return 0, tx.Error
 	}
-	return dexCount, nil
+	return int(dexCount), nil
 }
 
-func (g *gormDexRepository) CreateUserDexById(dexId int64, userId int64) (err error) {
+// [사용자 사건 수집 등록] 사건 id로 등록 post문
+func (g *gormDexRepository) CreateUserDexById(dexId int, userId int) (err error) {
 	// 1. 쿼리작성
 	// insert into userdex (dex_id, user_id)
 	// values (1, 1)
@@ -76,8 +80,8 @@ func (g *gormDexRepository) CreateUserDexById(dexId int64, userId int64) (err er
 	// 2. gorm로직
 	tx := g.db.Begin()
 	tx.Model(&entity.UserDex{}).Create(&entity.UserDex{
-		EventId: int(dexId),
-		UserId:  int(userId),
+		EventId: dexId,
+		UserId:  userId,
 	})
 	if tx.Error != nil {
 		tx.Rollback()
